% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0briskaR.R
\name{briskaRLoadInternProjection}
\alias{briskaRLoadInternProjection}
\alias{briskaRSetInternProjection}
\title{Load an internal working projection PROJ.4}
\usage{
briskaRSetInternProjection(proj = LAMBERT_93)
}
\arguments{
\item{proj}{A character string of projection arguments, must be in the PROJ.4 documentation}
}
\description{
Will load a projection as internal package working projection
}
