% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_FunctionalData.R
\name{brk_addFD}
\alias{brk_addFD}
\alias{brk_addFD2}
\title{Functional DATA}
\usage{
brk_addFD(sf, key, FUN, ...)

brk_addFD2(sf, keyConstraint, key, FUN, ...)
}
\arguments{
\item{sf}{object of class \link[sf]{sf}. See \link[sf]{sf} for details.}

\item{key}{the name of the new column, as strings or symbols}

\item{FUN}{the function to be applied to each element of \code{sf}.
In the case of functions like +, %*%, the function name must be
 backquoted or quoted. See lapply functions for details.}

\item{\dots}{optional arguments to FUN. See lapply functions for details.}

\item{keyConstraint}{character string. The reference of the column to be checked}
}
\value{
A \link[sf]{sf} object with addition functional data feature (or feature dynamic).
}
\description{
Add time series to sf objects
}
