% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_exposure.R
\name{brk_exposureMatch}
\alias{brk_exposureMatch}
\title{Add raster value to element of sf object}
\usage{
brk_exposureMatch(
  stackRaster_exposure,
  sf,
  stackTimeline,
  keyTime = "TIMELINE",
  key = "EXPOSURE"
)
}
\arguments{
\item{stackRaster_exposure}{The Raster* object}

\item{sf}{the sf object}

\item{stackTimeline}{sequence with matching to elements of RasterStack (length of stackTimeline
must be the same as length of a list of stackRaster_exposure)}

\item{keyTime}{name of the column to match exposure timeline from stackRaster_exposure}

\item{key}{name of the new column}
}
\description{
Add the raster value(s) of a Raster* object to element of sf object.
}
