% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_FunctionalData.R
\name{brk_FDtoDF}
\alias{brk_FDtoDF}
\alias{brk_FDtoDF_}
\alias{brk_FDtoDF_STICK}
\title{Convert list.column data.frame into scalar.column data.frame}
\usage{
brk_FDtoDF_(sf, key1, key2, id = NULL)

brk_FDtoDF(sf, key1, key2, id = NULL, keep = NULL)

brk_FDtoDF_STICK(sf, key1, keep)
}
\arguments{
\item{sf}{sf. An object of class \link[sf]{sf}}

\item{key1}{character string. The name of the column to select}

\item{key2}{icharacter string. The name of the column to select}

\item{id}{name of the replicate for the id. As to be of the same length as the number of row of the sf object}

\item{keep}{vector of column name to keep}
}
\description{
Convert data.frame with 2 column.list into data.frame with only column.scalar
}
