% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_exposure.R
\name{brk_newPoints}
\alias{brk_newPoints}
\title{Simulate new points on a specific sf object}
\usage{
brk_newPoints(sf, size = 100)
}
\arguments{
\item{sf}{object of class sf or sfc}

\item{size}{sample size(s) requested; either total size, or a numeric vector with sample sizes for each feature geometry. 
When sampling polygons, the returned sampling size may differ from the requested size, as the bounding box is
 sampled, and sampled points intersecting the polygon are returned.}
}
\description{
Simulate new points on a specific \link[sf]{sf} object. See \link[sf]{st_sample} for details.
}
