% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_FunctionalData.R
\name{brk_rbindLStoDF}
\alias{brk_rbindLStoDF}
\title{Combine list of data.frame by Rows}
\usage{
brk_rbindLStoDF(ls, id = NULL)
}
\arguments{
\item{ls}{A list of data.frame}

\item{id}{id to provide to each data.frame. Must be the length of the list}
}
\value{
Return a data.frame
}
\description{
Function used to filter functional data in \link[sf]{sf} objects.
}
