% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_sampling.R
\name{brk_sampling}
\alias{brk_sampling}
\title{Sampling based on distribution provided by Raster* object over a sf object}
\usage{
brk_sampling(rasterStack, sf, sizeSite = 1)
}
\arguments{
\item{rasterStack}{Raster* object (RasterLayer or RasterStack)}

\item{sf}{An sf object}

\item{sizeSite}{scalar or vector of the number of individual to sample per site. Default is 1.
If vector, as to be the length of Raster* object}
}
\description{
Sampling based on distribution provided by Raster* object 
(does not have to be summed to one) over an sf object used as mask
}
