% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brk_exposure.R
\name{brk_timeline}
\alias{brk_timeline}
\title{Add sequence to element of a \link[sf]{sf} object}
\usage{
brk_timeline(sf, key, from, to, by)
}
\arguments{
\item{sf}{object of class sf or sfc}

\item{key}{a character string used as name of the new column}

\item{from}{the starting value of the sequence. Of length 1 unless just from is supplied as an unnamed argument.}

\item{to}{the end (maximal) value of the sequence. Of length 1 unless just from is supplied as an unnamed argument.}

\item{by}{number: increment of the sequence.}
}
\description{
Add column with sequence for each row to a \link[sf]{sf} object
}
