% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briskaR.R
\docType{data}
\name{data_brk}
\alias{data_brk}
\alias{maize_65}
\alias{sfMaize65}
\alias{maize.emitted_pollen}
\alias{maize.proportion_pollen}
\alias{Hofmann_2009}
\alias{Lang_2004}
\alias{Precipitation}
\alias{temperatureGermany}
\alias{df_precipitation}
\title{Data set included in the package}
\usage{
data(maize_65)

data(sfMaize65)

data(maize.emitted_pollen)

data(maize.proportion_pollen)

data(Hofmann_2009)

data(Lang_2004)

data(Precipitation)

data(temperatureGermany)

data(df_precipitation)
}
\arguments{
\item{maize_65}{A SpatialPolygonsDataFrame of class \link[sp]{SpatialPolygonsDataFrame}  defining a patchy landscape}

\item{sfMaize65}{A set of MULTIPOLYGON of class \link[sf]{sf} defining a patchy landscape}

\item{maize.emitted_pollen}{A \link[base]{data.frame} of pollen emission}

\item{maize.proportion_pollen}{A \link[base]{data.frame} of pollen proportion}

\item{Hofmann_2009}{A \link[base]{data.frame} of pollen emission}

\item{Lang_2004}{A \link[base]{data.frame} of pollen emission}

\item{Precipitation}{\link[base]{data.frame} of daily Precipitation from 01/01/2003 to 31/12/2013}

\item{temperatureGermany}{A \link[base]{data.frame} of temperature over a year in south and north of germany}

\item{df_precipitation}{A \link[base]{data.frame} of daily Precipitation from 01/01/2003 to 31/12/2013}
}
\description{
Data set included in the package
}
