% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0briskaR.R
\name{GetInternProjection}
\alias{GetInternProjection}
\alias{briskaRGetInternProjection}
\title{Get the internal working projection PROJ.4}
\usage{
briskaRGetInternProjection()
}
\description{
Will print and return the internal projection of briskaR package
}
