% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{[,Individuals}
\alias{[,Individuals}
\alias{[,Individuals,numeric,missing,ANY-method}
\alias{[,Individuals,numeric,numeric,ANY-method}
\title{Get an individual information}
\usage{
\S4method{[}{Individuals,numeric,missing,ANY}(x, i)

\S4method{[}{Individuals,numeric,numeric,ANY}(x, i, j)
}
\arguments{
\item{x}{An \link{Individuals} object}

\item{i}{individual index}

\item{j}{time of information}
}
\value{
a data.frame with all Individuals slot(attributes) for an individual

a data.frame with all Individuals slot (attributes) for an individual at a specified time.
}
\description{
Get an Individual information

x[i] get individual i values from \link{Individuals} x.

Get an Individual information at a specified time

x[i,j] get individual i at time j values from \link{Individuals} x.
}
\examples{
data(maize_65)
# get individual 99 informations
str(maize.individuals[99])
data(maize_65)
#get individual 99 informations at time 30
maize.individuals[99,30]
}
