% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{print,Individuals-method}
\alias{print,Individuals-method}
\title{print Individuals informations}
\usage{
\S4method{print}{Individuals}(x, ...)
}
\arguments{
\item{x}{An \link{Individuals} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print Individuals informations
}
\examples{
data(maize_65)
print(maize.individuals)
}

