% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals-Methods.R
\docType{methods}
\name{show,Individuals-method}
\alias{show,Individuals-method}
\title{Show a summary of Individuals information}
\usage{
\S4method{show}{Individuals}(object)
}
\arguments{
\item{object}{An \link{Individuals} object}
}
\description{
print a summary of Individuals information (the 10 first individuals)
}
\examples{
data(maize_65)
maize.individuals
}
