% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{is_square_sf}
\alias{is_square_sf}
\title{Test if an \link[sf]{sf} is a square}
\usage{
is_square_sf(.sf, tolerance = 0.05)
}
\arguments{
\item{.sf}{and object of class \link[sf]{sf}}

\item{tolerance}{tolerance rate between both square side length}
}
\description{
Test if an sf is a square, with tolerance. Default is 5% tolerance.
}
