% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0briskaR.R
\docType{data}
\name{LAMBERT_93}
\alias{LAMBERT_93}
\title{LAMBERT_93}
\format{An object of class \code{character} of length 1.}
\usage{
LAMBERT_93
}
\description{
SIG projection Lambert_93 references "+init=epsg:2154" PROJ.4
}
\keyword{datasets}
