% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class-Landscape.R
\docType{class}
\name{Landscape}
\alias{Landscape}
\alias{Landscape-class}
\title{Class Landscape}
\description{
Class \code{Landscape} defines a landscape
}
\details{
\code{Landscape} objects can be created by calling of the allocator new("Landscape", ...), or (preferred) by calling to the function \link{simulateLandscape} or \link{loadLandscape}.
}
\section{Slots}{

\describe{
\item{\code{thelandscape}}{A SpatialPolygonsDataFrame}

\item{\code{xmin}}{Left x-axis coordinate (in meters)}

\item{\code{xmax}}{Right x-axis coordinate (in meters)}

\item{\code{ymin}}{Bottom y-axis coordinate (in meters)}

\item{\code{ymax}}{Top y-axis coordinate (in meters)}

\item{\code{n}}{Number of fields in the landscape}
}}

\seealso{
\link{simulateLandscape} , \link{loadLandscape} , \link{loadLandscapeSIG}
}
