% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{loadLandscape}
\alias{loadLandscape}
\title{Wrapper function : loadLandscape}
\usage{
loadLandscape(sp, data)
}
\arguments{
\item{sp}{a SpatialPolygons object designing the landscape}

\item{data}{a data.frame containing fields (polygons) information. Row names as fields ID, column names as sources | neutral | receptors (for a given field, the value is 1 for the type of the field (source or neutral or receptor), otherwise 0).}
}
\value{
A Landscape object
}
\description{
Wrapper function to create a Landscape object using SpatialPolygons and dataframe.
The SpatialPolygons object and the data.frame have to contain the same number of polygons
 and row (row ID is polygons ID).
}
\examples{
data(maize_65)
l<-loadLandscape(maize.landscape@thelandscape,maize.landscape@thelandscape@data)
plot(l)
}
