% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{loadLandscapeSIG}
\alias{loadLandscapeSIG}
\title{Create a Landscape object from SIG shapefile file}
\usage{
loadLandscapeSIG(dsn, layer, format = TRUE)
}
\arguments{
\item{dsn}{folder path to the source files}

\item{layer}{file name without extension}

\item{format}{only load data needed Landscpae-class (default TRUE)}
}
\value{
A Landscape object
}
\description{
Create a Landscape object from SIG shapefile. Shapefile has to contain a SpatialPolygonsDataFrame. Data in the data frame contain fields (polygons) information. Row names as fields ID, cols names as sources | neutral | receptors (for a given field, the value is 1 for the type of the field (source or neutral or receptor), otherwise 0).
}
\examples{
\dontrun{
land<-loadLandscapeSIG("/path/to/directory/","fileName")
plot(land)
}
}

