% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{print,Landscape-method}
\alias{print,Landscape-method}
\title{Print Method for Landscape}
\usage{
\S4method{print}{Landscape}(x, ...)
}
\arguments{
\item{x}{A Landscape object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Method for Landscape
}
\examples{
data(maize_65)
print(maize.landscape)
}
