% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\docType{methods}
\name{show,Landscape-method}
\alias{show,Landscape-method}
\title{Show Method for Landscape}
\usage{
\S4method{show}{Landscape}(object)
}
\arguments{
\item{object}{A Landscape class}
}
\description{
Show Method for Landscape
}
\examples{
data(maize_65)
maize.landscape 
}

