% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individuals_Methods.R
\name{loadIndividuals}
\alias{loadIndividuals}
\title{Wrapper function loadIndividuals}
\usage{
loadIndividuals(sf, data, timeline)
}
\arguments{
\item{sf}{a \link[sf]{sf} object}

\item{data}{a data.frame containing individuals attributes.
Rows numbers as individuals ID, columns names as dob (date of birth) | life_duration | toxic_threshold}

\item{timeline}{Vector of the time line}
}
\value{
an \code{Individuals} object
}
\description{
Wrapper function to create an individuals object using \link[sf]{sf} or \link[sp]{SpatialPoints} and data.frame.

The SpatialPoints object and the data.frame have to contain the same number of coordinates and rows.
}
