% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulPrecip.R
\name{loss_precipitation}
\alias{loss_precipitation}
\title{Loss precipitation function from a set of function
implemented in earlier version of briskaR}
\usage{
loss_precipitation(
  starttime = "16/07",
  endtime = "14/07",
  alpha = list(minalpha = 0.1, maxalpha = 0.95, covariate_threshold = 30, simulate =
    TRUE, covariate = NULL)
)
}
\arguments{
\item{starttime}{data of begining of simultation}

\item{endtime}{data of end of simultation}

\item{alpha}{list of parameter}
}
\description{
Loss precipitation function from a set of function
implemented in earlier version of briskaR
}
