\name{maize.emitted_pollen}
\alias{maize.emitted_pollen}
\docType{data}
\title{
  Pollen sources emission from maize crop.
}
\description{
  Pollen emission for the sources fields of the maize.landscape Landscape over 61 days.
}
\usage{data("maize.emitted_pollen")}
\format{
  A data frame with 184 observations (sources emission ID's by rownames) following by 61 values (time unit).
}
\examples{
data(maize.emitted_pollen)
}
\keyword{datasets}
