\name{maize.landscape}
\alias{maize.landscape}
\docType{data}
\title{
  Real Landscape maize_65 with attributs 
}
\description{
  The Landscape maize_65 with the distribution of neutral, sources, receptors fields.
}
%\usage{data("maize.landscape")}
\format{
  \code{Landscpae-class}
  }
\details{
  It contains 184 sources, 171 neutrals and  105 recpetors fields within the 460 fileds.
}
\examples{
data("maize.landscape")
plot(maize.landscape)
}
\keyword{datasets}
