\name{maize.proportion_pollen}
\alias{maize.proportion_pollen}
\docType{data}
\title{
Maize Proportion Pollen
}
\description{
The proportion of plants emitting pollen per day (during 12 days) was observed by Frédérique Angevin (Angevin et al. 2008).
}
\usage{data("maize.proportion_pollen")}
\format{
  The format is a vector of 12 days with proportions of plants emitting.
}
\source{
Frédérique Angevin (Angevin et al. 2008). 
}
\examples{
data(maize.proportion_pollen)
}
\keyword{datasets}
