\name{maize.toxicintensity61}
\alias{maize.toxicintensity61}
\docType{data}
\title{
  Pollen dispersion at time 61 over the maize.landscape from toxicIntensity
}
\description{
  Matrix of pollen dispersion at time 61 in the \code{maize.landscape} Landscape from \link{toxicIntensity} function.
}
%\usage{data("maize.toxicintensity61")}
\format{
  The format is a matrix 1024x1024 .
}
\examples{
data("maize.toxicintensity61")
}
\keyword{datasets}
