\name{maize_65}
\alias{maize_65}
\docType{data}
\title{
  Real Landscape
}
\description{
  Real Landscape from Hautes-Pyrénées (65) in South-West of France. The Landscape is an area of 5000x5000 meters near Rabastens-de-Bigorre taken in 2012.
}
\usage{data("maize_65")}
\format{
  \code{Landscape} class object.

  It contains 315 maize sources fields within 460 fields.
}
\source{
  Open data come from https://www.data.gouv.fr/fr/datasets/registre-parcellaire-graphique-2012-contours-des-ilots-culturaux-et-leur-groupe-de-cultures-majorita/
}
\examples{
data(maize_65)
}
\keyword{datasets}
