% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{nlayers}
\alias{nlayers}
\title{A set of function imported from `raster` package}
\description{
Re-exported \code{nlayers} method from `raster`
package to avoid failing when working on `raster` objects without having 
`raster` loaded first.
}
\keyword{internal}
