% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot from raster and sp (sf)}
\description{
Re-exported \code{plot} method from `raster`, `sp` and `sf`
packages to avoid failing when trying to plot `raster` or `sf` objects without
 having `raster` or `sf` loaded first.
}
\note{
For \link[sf]{sf} objects, the package `ggplot2` include the function `geom_sf`
}
\keyword{internal}
