\name{Precipitation}
\alias{Precipitation}
\docType{data}
\title{
Precipitation data
}
\description{
  Precipitation Data from simulation
}
\usage{data("Precipitation")}
\format{
  A data frame with 4018 observations on the following 4 variables.
  \describe{
    \item{\code{AN}}{Year}
    \item{\code{MOIS}}{Month}
    \item{\code{JOUR}}{Day}
    \item{\code{RR}}{Precipitation ratio}
  }
}
\details{
  Precipitation over 10 years from 2003 to 2013
}
\source{
  From simulations of the Stochastic Weather Generator developed by Allard and Bourotte (2014).
}
\examples{
data(Precipitation)
}
\keyword{datasets}
