% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{rasterToPoints}
\alias{rasterToPoints}
\title{A set of function imported from `raster` package}
\description{
Raster to points conversion. Cells with NA are not converted. A function
can be used to select a subset of the raster cells (by their values).
}
\keyword{internal}
