% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulPrecip.R
\name{simul.precipitation}
\alias{simul.precipitation}
\title{Simulate precipitation between two dates}
\usage{
simul.precipitation(starttime = "15/07", endtime = "15/09", data = NULL)
}
\arguments{
\item{starttime}{*string*. Date shape: \code{"mm/yy"}}

\item{endtime}{*string*. Date shape: \code{"mm/yy"}}

\item{data}{*data.frame*. default is \code{NULL}}
}
\value{
an array of length of the period between the two dates included
}
\description{
Will evaluate parameters from data and simulate precipitation between the two dates.
}
\author{
Jean-Francois Rey
}
