% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{st_geometry}
\alias{st_geometry}
\title{Get, set, or replace geometry from an `sf`` object}
\description{
Compute Euclidian or great circle distance
 between pairs of geometries; compute, the area or the
 length of a set of geometries.
}
\keyword{internal}
