% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{st_length}
\alias{st_length}
\title{Compute geometric measurements}
\description{
Compute Euclidian or great circle distance
 between pairs of geometries; compute, the area or the
 length of a set of geometries.
}
\keyword{internal}
