% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{st_sample}
\alias{st_sample}
\title{sample points on or in (sets of) spatial features}
\description{
Sample points on or in (sets of) spatial features.
By default, returns a pre-specified number of points that is equal to size
(if type = "random") or an approximation of size (for other sampling types).
}
\keyword{internal}
