% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{st_squared_geometry}
\alias{st_squared_geometry}
\title{Add squared frame polygon}
\usage{
st_squared_geometry(list_sf, buffer_dist = NULL)
}
\arguments{
\item{list_sf}{A list of objects of class \link[sf]{sf}}

\item{buffer_dist}{numeric; buffer distance for all, or for each of the elements in x;
in case dist is a units object, it should be convertible to arc_degree
 if x has geographic coordinates, and to st_crs(x)$units otherwise. 
 See function `sf_buffer` from package `sf` for details}
}
\description{
Return a square frame surrounding a list of sf
}
